Rust PROBE frontend (#20)
* rust PROBE frontend

* Added logging facility

* Update .envrc

Update the .envrc for probe_frontend to automatically run make on
libprobe, as well as export the __PROBE_LIB environment variable so that
the frontend can find the libprobe.so

* Documentation and code cleanup

* Apply suggestions from code review

Co-authored-by: Sam Grayson <sam@samgrayson.me>

* More concise Display impls

Removes a lot of details that are usually abstracted by libraries when
printing for humans. In contrast with some of the code review comments I
left mtime and errno field because I feel those are useful for human
readers; even if all a human is likley to do with errno is "was it
zero?"

* Refactor Ops to use CString instead of OsString

* Added additional cargo devtools to probe_frontend devShell

* Major refactor

This commit addresses serveral points raised durring the last
code-review (and a few that weren't):

- Restructured arena.rs to read top-down.
- Refactored child process code to use std::process instead of
subprocess crate.
- Cleaned up arena pointer resolition to use less unsafe code.
- [INFO] Better log messages.

* version 0.2.0

* fix cargo fmt/clippy

* Added pygen sanity check to flake checks

* :sparkles: documentation :sparkles:

* Update README.md

* typos

* Fixed typo

Co-authored-by: Sam Grayson <sam@samgrayson.me>

* code review improvements

* improved pygen code

* improved rust struct generation

the structs now get converted to camel case at the rust level instead of
the python level meaning python doesn't have to do any type name
translation. Also strips out redundant field prefixes from the rust
structs (which also solves the issue for python).

* pygen custom `@property`s

* improved arg parsing

* improved readme

* cleanup and documentation

* `probe record` warn on non-zero exit code

* disable overzelous warning when running under gdb

* switched gdb to internal binary shim

* proc macros return compile_error!() instead of panicking

* added runtime check for arena instantiation

* fixed README.md

* fixed `--gdb`

* probe_log decoding now matches parse_probe_log.py

* build probe cli with static linking

building the probe cli with `cargo build` inside the devShell will still
produce a dynamically linked binary, but building with `nix build
.#probe-cli` should produce a statically linked binary

* update frontend code to work with main-branch libprobe

* fixed ruff errors in generated python code

* added LICENSE

* added probe-py derivation to output generated python module

* revised probe-py derivation

---------

Co-authored-by: Sam Grayson <sam@samgrayson.me>