Remove argc count (#123)
This PR removes extraneous counts for the number of arguments and environment variables, where it is neither necessary nor helpful. At one point, we thought the Rust parser would have an easier time reading the `char**` if it knew the number of elements, but we figured out how to do that without knowing the number of elements.

The number of elements is still helpful for performance but not necessary when executing `arena_copy_argv`, but those calls are split into two cases:

- In use-cases where we _already have the number of elements on-hand_, we should pass that to `arena_copy_argv`;
- In cases where we _computed the number specifically for `arena_copy_argv`_ were eliminated. Computing it in calling code is no more efficient than computing it in `arena_copy_argv`.

In both cases, we no longer need to take the number of elements as a pointer, since we don't need to use it later on; passing-by-value will do.