Initialize libprobe in ELF constructor (#121)
Previously, PROBE becomes initialized on the first intercepted call. The ELF constructor section is a more reliable place to initialize because some programs might not make relevant libcalls that would get intercepted by PROBE, `/bin/true` for example. Therefore, we add `/bin/true` as a test case, which would fail before this PR by failing `probe validate`, specifically the point where we check that every `ExecOp` is associated with an `InitExecEpochOp`.

This initialization is the most logical place to store the command-line arguments, current executable, and environment variables, since the child process is more likely to execute the ELF constructor than the parent is to call a dynamic `exec*` symbol. For example, a program that makes raw system calls would execute the former but not the latter.

This brings up the question whether to emit `InitProcessOp` and `InitExecEpochOp` or just the latter. Previously, `libprobe` emits `InitProcessOp` if `libprobe` thinks that this is the first exec epoch of that process. That does not correspond to an observed event, rather it is an inferred event based on the observed initialization of the exec epoch. Therefore, it is more appropriately left to the analysis code to insert and infer it wishes, and `libprobe` should only record observed events (no inferred). This PR only emits `InitExecEpochOp`.