Include Path in CloseOp (#132)
When we close a file descriptor, we should record which path got closed. The Path struct notably includes the mod time of the file, which determines the version of the path. Including that will help us check that our analysis tracked file descriptors and versions correctly. In most cases, it is only a check, but in a few cases, it actually gives us information that we wouldn't otherwise know.