Make pipe and fifo calls work (#143)
`mkdir` and `mkfifo` are similar operations that create a "special" file, a directory file and a FIFO file respectively. They have therefore been given a combined opcode with an enum to distinguish the two types. `pipe` is similar to `mkfifo` followed be two `open`s (one for the read; one for the write side of the pipe).